LayoutTests/http/tests/appcache/abort-cache-onchecking-manifest-404-expected.txt

CONSOLE MESSAGE: line 1: ApplicationCache is deprecated. Please use ServiceWorkers instead.
CONSOLE MESSAGE: Application Cache manifest could not be fetched, because the manifest had a 404 response.
This tests that download process was aborted after checking event and before update failed.
SUCCESS
